modified with script

+++ push.sh
@@ -13,11 +13,20 @@
 #	* Sat Dec 16 2017 Vinu K <kevy.vinu@gmail.com
 #	- Original
 
-if [[ -d _includes ]]; then
-	echo "Last Updated $(date)" > _includes/time.html
-	git add .
-	git commit -m "modified with script"
-	git push origin master
-else
-	echo "No such file like time.html"
-fi
+git status
+read -r -p "Are you sure want to commit? [y/N] " response
+case "$response" in
+    [yY][eE][sS]|[yY])
+	    if [[ -d _includes ]]; then
+		    echo "Last Updated $(date)" > _includes/time.html
+		    git add .
+		    git commit -m "modified with script"
+		    git push origin master
+	    else
+		    echo "No such file like time.html"
+	    fi
+	    ;;
+    *)
+	    echo "Aborted"
+	    ;;
+esac
